+++ src/components/App-components/Modals/ConfirmationModal.module.css
@@ -5,6 +5,7 @@
   width: 100%;
   height: 100%;
   background-color: rgba(0, 0, 0, 0.5);
+  z-index: 50;
 }
 
 .buttonContainer {
