+++ src/components/App-components/Modals/ModalFormRequest.jsx
@@ -1,5 +1,6 @@
 import PropTypes from "prop-types";
-import { Button } from "@mui/material";
+import { Modal, Button } from "@mui/material";
+
 import { useFormik } from "formik";
 import * as Yup from "yup";
 // eslint-disable-next-line import/no-unresolved
@@ -20,6 +21,7 @@ export default function ModalFormRequest({
   handleOpenConfirmationModal,
   requestId,
   handleCreateOrUpdateRequest,
+  openModalAboutRequest,
 }) {
   const theme = useTheme();
 
@@ -118,11 +120,12 @@ export default function ModalFormRequest({
     formik.setValues(newData);
   };
   // function to check if a value has been changed in the form. If don't : close directly modal
-  const isDataChanged = requestId
-    ? Object.keys(formik.values).some(
-        (key) => formik.values[key] !== requestData[key]
-      )
-    : Object.values(formik.values).some((value) => value !== "");
+  const isDataChanged =
+    requestId && requestData
+      ? Object.keys(formik.values).some(
+          (key) => formik.values[key] !== requestData[key]
+        )
+      : Object.values(formik.values).some((value) => value !== "");
 
   // Update form values when initialValues change
   useEffect(() => {
@@ -136,118 +139,152 @@ export default function ModalFormRequest({
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [requestData]);
+  // modal style
+  const style = {
+    position: "absolute",
+    top: "50%",
+    left: "50%",
+    transform: "translate(-50%, -50%)",
+    height: 700,
+    bgcolor: "#292929",
+    borderRadius: "10px",
+    boxShadow: 24,
+    p: 4,
+    display: "flex",
+    justifyContent: "center",
+    width: {
+      sm: "400px",
+      md: "450px",
+      lg: "500px",
+      xl: "550px",
+    },
+  };
 
   return (
-    <div className={styles.formStyle}>
-      <button
-        className={styles.buttonClose}
-        onClick={() => {
-          if (isDataChanged) {
-            handleOpenConfirmationModal();
-          } else {
-            handleClose();
-          }
-        }}
-        type="button"
-      >
-        X
-      </button>
-      <form onSubmit={formik.handleSubmit} className={styles.form}>
-        <h1>{modalTitle} </h1>
-        <div className={styles.input}>
-          <div className={styles.inputStyle}>
-            <TextInput
-              label="Nom de la PR"
-              type="text"
-              id="title"
-              placeholder="Nommez votre PR"
-              onChange={formik.handleChange}
-              onBlur={formik.handleBlur}
-              value={formik.values.title}
-              readOnly={false}
-            />
-            {formik.touched.title && formik.errors.title ? (
-              <TooltipIcon
-                tooltip={formik.errors.title}
-                top="0"
-                left="130px"
-                color="var(--error)"
+    <Modal
+      sx={style}
+      open={openModalAboutRequest}
+      onClose={() => {
+        if (isDataChanged) {
+          handleOpenConfirmationModal();
+        } else {
+          handleClose();
+        }
+      }}
+      aria-labelledby="modal-modal-title"
+      aria-describedby="modal-modal-description"
+    >
+      <div className={styles.formStyle}>
+        <button
+          className={styles.buttonClose}
+          onClick={() => {
+            if (isDataChanged) {
+              handleOpenConfirmationModal();
+            } else {
+              handleClose();
+            }
+          }}
+          type="button"
+        >
+          X
+        </button>
+        <form onSubmit={formik.handleSubmit} className={styles.form}>
+          <h1>{modalTitle} </h1>
+          <div className={styles.input}>
+            <div className={styles.inputStyle}>
+              <TextInput
+                label="Nom de la PR"
+                type="text"
+                id="title"
+                placeholder="Nommez votre PR"
+                onChange={formik.handleChange}
+                onBlur={formik.handleBlur}
+                value={formik.values.title}
+                readOnly={false}
               />
-            ) : null}
-          </div>
-          <div className={styles.inputStyle}>
-            <TextArea
-              label="Description"
-              type="text"
-              id="description"
-              placeholder="Décrivez votre PR"
-              onChange={formik.handleChange}
-              onBlur={formik.handleBlur}
-              value={formik.values.description}
-            />
-            {formik.touched.description && formik.errors.description ? (
-              <TooltipIcon
-                tooltip={formik.errors.description}
-                top="0"
-                left="110px"
-                color="var(--error)"
+              {formik.touched.title && formik.errors.title ? (
+                <TooltipIcon
+                  tooltip={formik.errors.title}
+                  top="0"
+                  left="130px"
+                  color="var(--error)"
+                />
+              ) : null}
+            </div>
+            <div className={styles.inputStyle}>
+              <TextArea
+                label="Description"
+                type="text"
+                id="description"
+                placeholder="Décrivez votre PR"
+                onChange={formik.handleChange}
+                onBlur={formik.handleBlur}
+                value={formik.values.description}
               />
-            ) : null}
-          </div>
-          <div className={styles.inputStyle}>
-            <TextInput
-              label="Lien Trello"
-              type="text"
-              id="trello"
-              placeholder="Insérez le lien vers Trello"
-              onChange={formik.handleChange}
-              onBlur={formik.handleBlur}
-              value={formik.values.trello}
-            />
-            {formik.touched.trello && formik.errors.trello ? (
-              <TooltipIcon
-                tooltip={formik.errors.trello}
-                top="0"
-                left="100px"
-                color="var(--error)"
+              {formik.touched.description && formik.errors.description ? (
+                <TooltipIcon
+                  tooltip={formik.errors.description}
+                  top="0"
+                  left="110px"
+                  color="var(--error)"
+                />
+              ) : null}
+            </div>
+            <div className={styles.inputStyle}>
+              <TextInput
+                label="Lien Trello"
+                type="text"
+                id="trello"
+                placeholder="Insérez le lien vers Trello"
+                onChange={formik.handleChange}
+                onBlur={formik.handleBlur}
+                value={formik.values.trello}
               />
-            ) : null}
-          </div>
-          <div className={styles.inputStyle}>
-            <TextInput
-              label="Lien Github"
-              type="text"
-              id="github"
-              placeholder="Insérez le lien vers Github"
-              onChange={formik.handleChange}
-              onBlur={formik.handleBlur}
-              value={formik.values.github}
-            />
-            {formik.touched.github && formik.errors.github ? (
-              <TooltipIcon
-                tooltip={formik.errors.github}
-                top="0"
-                left="110px"
-                color="var(--error)"
+              {formik.touched.trello && formik.errors.trello ? (
+                <TooltipIcon
+                  tooltip={formik.errors.trello}
+                  top="0"
+                  left="100px"
+                  color="var(--error)"
+                />
+              ) : null}
+            </div>
+            <div className={styles.inputStyle}>
+              <TextInput
+                label="Lien Github"
+                type="text"
+                id="github"
+                placeholder="Insérez le lien vers Github"
+                onChange={formik.handleChange}
+                onBlur={formik.handleBlur}
+                value={formik.values.github}
               />
-            ) : null}
+              {formik.touched.github && formik.errors.github ? (
+                <TooltipIcon
+                  tooltip={formik.errors.github}
+                  top="0"
+                  left="110px"
+                  color="var(--error)"
+                />
+              ) : null}
+            </div>
           </div>
-        </div>
-        <div className={styles.button}>
-          <Button
-            variant="contained"
-            onClick={handleSaveClick}
-            sx={{
-              width: ["100%", "100%", "40%"],
-              background: theme.palette.button.main,
-              fontFamily: theme.typography.fontFamily,
-            }}
-          >
-            {modalTitle}
-          </Button>
-        </div>
-      </form>
-    </div>
+          <div className={styles.button}>
+            <Button
+              variant="contained"
+              onClick={handleSaveClick}
+              sx={{
+                width: ["100%", "100%", "40%"],
+                background: theme.palette.button.main,
+                fontFamily: theme.typography.fontFamily,
+              }}
+            >
+              {modalTitle}
+            </Button>
+          </div>
+        </form>
+      </div>
+    </Modal>
   );
 }
 ModalFormRequest.propTypes = {
@@ -256,6 +293,7 @@ ModalFormRequest.propTypes = {
   handleOpenConfirmationModal: PropTypes.func.isRequired,
   requestId: PropTypes.number,
   handleCreateOrUpdateRequest: PropTypes.func.isRequired,
+  openModalAboutRequest: PropTypes.bool.isRequired,
 };
 ModalFormRequest.defaultProps = {
   requestId: null,
